src/llm-compiler.ts
import { BaseLanguageModel } from "@langchain/core/language_models/base";
import { Tool } from "@langchain/core/tools";
import { LLMCompilerError, SubtaskGenerationError, CodeGenerationError, CodeAnalysisError, ExecutionPlanError } from "./errors";
import { DAG } from "./dag";
import { CompilerState, Task, CodeBlock } from "./types";

export class LLMCompiler {
  private llm: BaseLanguageModel;
  private tools: Tool[];

  constructor(llm: BaseLanguageModel, tools: Tool[]) {
    this.llm = llm;
    this.tools = tools;
  }

  async compile(task: string): Promise<CompilerState> {
    try {
      console.log(`Starting compilation for task: ${task}`);
      const subtasks = await this.generateSubtasks(task);
      const dag = this.buildDAG(subtasks);
      const codeBlocks = await this.generateCode(subtasks);
      const analysis = await this.analyzeCode(codeBlocks);
      const plan = await this.generateExecutionPlan(dag);
      const integratedCode = this.integrateCode(codeBlocks);

      console.log('Compilation completed successfully');
      return {
        task,
        subtasks,
        dag,
        code_blocks: codeBlocks,
        analysis,
        plan,
        integrated_code: integratedCode,
        execution_output: "Execution not implemented in this version"
      };
    } catch (error) {
      console.error('Error during compilation:', error);
      if (error instanceof Error) {
        throw new LLMCompilerError(`LLM Compiler failed: ${error.message}`);
      }
      throw new LLMCompilerError('An unknown error occurred during LLM compilation');
    }
  }

  private async generateSubtasks(task: string): Promise<Task[]> {
    try {
      console.log('Generating subtasks...');
      const subtasksResult = await this.llm.predict(`Generate subtasks for the following task: ${task}`);
      return subtasksResult.split('\n')
        .map(subtask => subtask.trim())
        .filter(subtask => subtask.length > 0)
        .map((subtask, index) => ({
          id: `task_${index + 1}`,
          description: subtask
        }));
    } catch (error) {
      throw new SubtaskGenerationError((error as Error).message);
    }
  }

  private buildDAG(subtasks: Task[]): DAG {
    const dag = new DAG();
    subtasks.forEach(task => dag.addNode(task.id, task));
    // Here you can implement logic to add edges between tasks based on their dependencies
    return dag;
  }

  private async generateCode(subtasks: Task[]): Promise<CodeBlock[]> {
    try {
      console.log('Generating code...');
      return await Promise.all(subtasks.map(async (task) => {
        const codeResult = await this.llm.predict(`Generate code for the following task: ${task.description}`);
        return {
          taskId: task.id,
          code: codeResult
        };
      }));
    } catch (error) {
      throw new CodeGenerationError((error as Error).message);
    }
  }

  private async analyzeCode(codeBlocks: CodeBlock[]): Promise<string> {
    try {
      console.log('Analyzing code...');
      const combinedCode = codeBlocks.map(block => block.code).join('\n\n');
      return await this.llm.predict(`Analyze the following code:\n${combinedCode}`);
    } catch (error) {
      throw new CodeAnalysisError((error as Error).message);
    }
  }

  private async generateExecutionPlan(dag: DAG): Promise<string> {
    try {
      console.log('Generating execution plan...');
      const sortedTasks = dag.topologicalSort();
      return await this.llm.predict(`Generate an execution plan for the following task order: ${sortedTasks.join(', ')}`);
    } catch (error) {
      throw new ExecutionPlanError((error as Error).message);
    }
  }

  private integrateCode(codeBlocks: CodeBlock[]): string {
    return codeBlocks.map(block => `// Task: ${block.taskId}\n${block.code}`).join('\n\n');
  }
}